+++ OntologiaVacuidad.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Models", "Models\Models.csp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Models.Tests", "Models.Tests\Models.Tests.csproj", "{FE8CAEAA-2F97-4573-B1EF-2764F49056A1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DomainLogic", "DomainLogic\DomainLogic.csproj", "{36156E55-C37E-4886-9031-D4DFCAA2534A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,18 @@ Global
 		{FE8CAEAA-2F97-4573-B1EF-2764F49056A1}.Release|x64.Build.0 = Release|Any CPU
 		{FE8CAEAA-2F97-4573-B1EF-2764F49056A1}.Release|x86.ActiveCfg = Release|Any CPU
 		{FE8CAEAA-2F97-4573-B1EF-2764F49056A1}.Release|x86.Build.0 = Release|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Debug|x64.Build.0 = Debug|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Debug|x86.Build.0 = Debug|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Release|x64.ActiveCfg = Release|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Release|x64.Build.0 = Release|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Release|x86.ActiveCfg = Release|Any CPU
+		{36156E55-C37E-4886-9031-D4DFCAA2534A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
